recon/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
strawberry-graphql[fastapi]==0.219.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Text Similarity
python-Levenshtein==0.25.0
fuzzywuzzy==0.18.0

# Date utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0